New consensus topic COMMIT
